+++ ios/ViroRenderer/ViroKit.framework/Headers/VROVideoDelegateInternal.h
@@ -21,7 +21,7 @@ public:
     virtual void videoWillBuffer() = 0;
     virtual void videoDidBuffer() = 0;
     virtual void videoDidFinish() = 0;
-    virtual void onVideoUpdatedTime(int currentTimeInSeconds, int totalTimeInSeconds) = 0;
+    virtual void onVideoUpdatedTime(float currentTimeInSeconds, float totalTimeInSeconds) = 0;
     virtual void videoDidFail(std::string error) = 0;
 };
